Replace deprecated source_permissions parameter
The parameter has been deprecated by Puppetlabs and will be removed in
Puppet6. In order to ensure that the module keeps working in the future
the functionality of the parameter had to be replaced.

The only easy way to solve this, is by using an exec statement that uses
rsync to sync the files from the source folder