Fixes #374 - Revocation command update and crl renew
An issue was raised informing that the revocation command is incorrect.
This was diagnosed to indeed be the case. As the `$name` variable in
context of `revoke.pp` does not evalute to `server name` but instead
`client name`. The exec for the crl renew was updated to clarify which server
it's done for and to prevent duplicate `exec` resource names.

`catch_changes` in the acceptance test was taken out because a crl renew
is triggrered which is seen as a change.